+++ src/main/scala/com/stevens/Queryable.scala
@@ -1,5 +1,5 @@
 package com.stevens
 
 trait Queryable {
-  def findNearestNeighbor(point: Array[Float]): LabeledPoint
+  def findNearestNeighbor(point: Array[Int]): LabeledPoint
 }
